move instance collection files
