+++ src/activities/instances/instanceCollections/CollectedObject.js
@@ -1,4 +1,4 @@
-import { Entity } from '../../es6/Entity.js';
+import { Entity } from '../../../es6/Entity.js';
 
 export class CollectedObjectEntity extends Entity {
 	sortOrder() {
