+++ test/activities/instances/instanceCollections/data/CollectedObject.js
